% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{cde}
\alias{cde}
\alias{cde_gram_per_hour}
\title{CDE}
\usage{
cde(mass, mac, gwp, fgf, temperature = 293.15, pressure = 101.325)

cde_gram_per_hour(
  mass,
  mac,
  gwp,
  fgf,
  temperature = 293.15,
  pressure = 101.325
)
}
\arguments{
\item{mass}{\code{double}, molare mass in g/mol.}

\item{mac}{\code{double}, minimum alveolar concentration (as ratio, e.g. for 2 \%,
0.02).}

\item{gwp}{\code{double}, global warming potential.}

\item{fgf}{\code{double}, fresh gas flow in litre per minute.}

\item{temperature}{\code{double} in Kelvin.}

\item{pressure}{\code{double} in kPa.}
}
\value{
\code{double}, carbon dioxide equivalent in gram per hour.
}
\description{
Carbon dioxide equivalent.
}
\examples{
# sevofluran, gwp20
cde(200, 0.02, 349, 0.5)

# same using package provided data
data(ianes)
sevo <- subset(ianes, Name == "Sevoflurane")
# for 20 years
cde(sevo$MW, sevo$MAC, sevo$GWP20, 0.5)
# for 100 years
cde(sevo$MW, sevo$MAC, sevo$GWP100, 0.5)
}
