% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuels}
\alias{fuels}
\title{Fuels}
\format{
A \code{data.frame} with 2 rows and 2 columns:
\describe{
\item{Name}{\code{character}, name.}
\item{CO2}{\code{double}, gram CO2 per litre fuel.}
}
}
\usage{
fuels
}
\description{
A dataset with carbon dioxid data of the some fuels.
}
\references{
Umweltbundesamt. "Ratgeber: Sprit sparen und mobil sein" (2009).
\url{https://www.umweltbundesamt.de/sites/default/files/medien/publikation/long/3705.pdf}
accessed 2021-11-11.
}
\keyword{datasets}
