% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{gram_per_hour}
\alias{gram_per_hour}
\title{Gram of Volatile Anaesthetic per Hour}
\usage{
gram_per_hour(mass, mac, fgf, temperature = 293.15, pressure = 101.325)
}
\arguments{
\item{mass}{\code{double}, molare mass in g/mol.}

\item{mac}{\code{double}, minimum alveolar concentration (as ratio, e.g. for 2 \%,
0.02).}

\item{fgf}{\code{double}, fresh gas flow in litre per minute.}

\item{temperature}{\code{double} in Kelvin.}

\item{pressure}{\code{double} in kPa.}
}
\value{
\code{double}, gram of volatile anaesthetic per hour.
}
\description{
Gram of Volatile Anaesthetic per Hour
}
\examples{
gram_per_hour(200, 0.02, 0.5)
}
