% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ianes}
\alias{ianes}
\title{Inhalational Anesthetics}
\format{
A \code{data.frame} with 3 rows and 7 columns:
\describe{
\item{Name}{\code{character}, name.}
\item{MAC}{\code{double}, minimum alveolar concentration.}
\item{MW}{\code{double}, molecular weight.}
\item{Density}{\code{double}, density.}
\item{GWP20}{\code{double}, global warming potential for 20 years.}
\item{GWP100}{\code{double}, global warming potential for 100 years.}
\item{GWP500}{\code{double}, global warming potential for 500 years.}
\item{Unit}{\code{numeric}, mili litre per unit/bottle.}
\item{Color}{\code{character}, color code.}
}
}
\usage{
ianes
}
\description{
A dataset of physical properties of the some inhalational anesthetics.
}
\references{
Ryan, Susan M., and Claus J. Nielsen.
"Global warming potential of inhaled anesthetics: application to clinical use."
Anesthesia & Analgesia 111.1 (2010): 92-98.
\doi{10.1213/ane.0b013e3181e058d7}
}
\keyword{datasets}
