% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{litre_per_hour}
\alias{litre_per_hour}
\title{Litre of Volatile Anaesthetic per Hour}
\usage{
litre_per_hour(mac, fgf)
}
\arguments{
\item{mac}{\code{double}, minimum alveolar concentration (as ratio, e.g. for 2 \%,
0.02).}

\item{fgf}{\code{double}, fresh gas flow in litre per minute.}
}
\value{
\code{double}, litre of volatile anaesthetic per hour.
}
\description{
Litre of Volatile Anaesthetic per Hour
}
\examples{
litre_per_hour(0.02, 0.5)
}
