% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ianes}
\alias{ianes}
\title{Inhalational Anesthetics}
\format{
A \code{data.frame} with 3 rows and 9 columns:
\describe{
\item{Name}{\code{character}, name.}
\item{MAC}{\code{double}, minimum alveolar concentration.}
\item{MW}{\code{double}, molecular weight.}
\item{Density}{\code{double}, density.}
\item{GWP20}{\code{double}, global warming potential for 20 years.}
\item{GWP100}{\code{double}, global warming potential for 100 years.}
\item{GWP500}{\code{double}, global warming potential for 500 years.}
\item{Unit}{\code{numeric}, mili litre per unit/bottle.}
\item{Color}{\code{character}, color code.}
}
}
\usage{
ianes
}
\description{
A dataset of physical properties of the some inhalational anesthetics.
}
\references{
Andersen, Nielsen, Wallington, Karpichev, and Sander
"Assessing the Impact on Global Climate from General Anesthetic Gases"
Anesthesia & Analgesia 114.5 (2012): 1081-1085
\doi{10.1213/ANE.0b013e31824d6150}

Andersen, Nielsen, and Sherman
"The Global Warming Potentials for Anesthetic Gas Sevoflurane Need
Significant Corrections"
Environ. Sci. Technol. 2021, 55, 15, 10189–10191
\doi{10.1021/acs.est.1c02573}
}
\keyword{datasets}
